Decompose Reborn sandbox broker wiring
